Fix compiler warning.

+++ trunk/src/jsformatter.cpp
@@ -80,7 +80,7 @@ DLLAPI const char *JSFGetDebugOutput(JSFormatter *jsf)
 	return ((RealJSFormatter *)jsf)->GetDebugOutput();
 }
 
-DLLAPI int JSFGetFormattedLine(JSFormatter *jsf, int originalLine)
+DLLAPI long JSFGetFormattedLine(JSFormatter *jsf, long originalLine)
 {
 	return ((RealJSFormatter *)jsf)->GetFormattedLine(originalLine);
 }
